requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1;python_version<"3.9"
bleach==4.1.0
Django==4.1.4
django-ckeditor==6.5.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-rest-multiple-models==2.1.3
django-solo==2.0.0
django-summernote==0.8.20.0
djangorestframework==3.14.0
gunicorn==20.1.0
packaging==23.0
Pillow==9.4.0
psycopg2-binary==2.9.5
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
webencodings==0.5.1
whitenoise==6.3.0
